urku-steps/backend/requirements.txt
Flask
requests
Pillow
folium
OWSLib
matplotlib
